backend/requirements.txt
aiohappyeyeballs
typing_extensions
anyio
asgiref
attrs
cachetools
certifi
charset-normalizer
distro
Django
djangorestframework
frozenlist
google-api-core
google-auth
google-cloud-texttospeech
google-cloud-vision
googleapis-common-protos
grpcio
grpcio-status
h11
httpcore
httpx
idna
jiter
multidict
openai
pillow
pip
propcache
proto-plus
protobuf
pyasn1
pyasn1_modules
pydantic
pydantic_core
requests
rsa
sniffio
sqlparse
tqdm
typing_extensions
urllib3
yarl